components/inventory_management/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
bson==0.5.10
certifi==2025.8.3
cffi==1.17.1
colorama==0.4.6
coverage==7.10.6
cryptography==45.0.7
dnspython==2.7.0
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jose==1.0.0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.14.1
pyparsing==3.2.3
pytest==8.4.2
pytest-cov==6.3.0
pytest-mock==3.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
PyYAML==6.0.2
rsa==4.9.1
-e git+https://github.com/arunkalyan12/Inventory-Management.git@edffca21977a0283263c604b6eebf192347bf166#egg=shared_utils&subdirectory=components\shared_utils
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.15.0
